update CachedVocabularyRepository: add GetByTopicId

+++ src/Features/Vocabulary/InterfaceAdapters/CachedVocabularyRepositoryInterface.cs
@@ -5,5 +5,5 @@ namespace Core.Features.Vocabulary.InterfaceAdapters;
 public interface CachedVocabularyRepositoryInterface {
     public Task<GetListVocabulariesResult> GetAll();
     public Task<GetVocabularyResult> Get(long id);
-    public Task<GetListVocabulariesResult> GetByTopicId(long topicId);
+    public Task<GetListVocabulariesResult> GetByTopicId(int topicId);
 }
